pyproject.toml
[tool.poetry]
name = "geo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12,<3.13"
numpy = ">=1.21,<2.0"
scikit-learn = "^1.7.1"
matplotlib = "^3.10.3"
tqdm = "^4.67.1"
polars = "^1.31.0"
torch-geometric = "2.6.1"
requests = "^2.32.4"
seaborn = "^0.13.2"
torch = {version = "^2.9.0.dev20250725+cu129", allow-prereleases = true, source = "pytorch-nightly-cu129"}
torchvision = {version = "^0.24.0.dev20250725+cu129", allow-prereleases = true, source = "pytorch-nightly-cu129"}
torchaudio = {version = "^2.8.0.dev20250725+cu129", allow-prereleases = true, source = "pytorch-nightly-cu129"}
pyarrow = "^21.0.0"
faiss-gpu-cu12 = "^1.11.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
tensorboard = "^2.20.0"
black = "^25.1.0"
pytest = "^8.4.1"
ipykernel = "^6.30.0"


[[tool.poetry.source]]
name = "pytorch-nightly-cu129"
url = "https://download.pytorch.org/whl/nightly/cu129"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
